move Gnu-specific feature '--print-search-dirs' to GnuLikeCompiler
